amelioration de l'espace admin

+++ src/main/java/com/gestionregion/gestionregionfront/security/models/Role.java
@@ -1,17 +1,22 @@
 package com.gestionregion.gestionregionfront.security.models;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Entity
 @Table(name = "roles")
 public class Role {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
 
   @Enumerated(EnumType.STRING)
   @Column(length = 20)
   private ERole name;
+
 
   public Role() {
 
@@ -21,11 +26,11 @@ public class Role {
     this.name = name;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
